review fixes: block_headers_status_notifier, block_headers_status_watcher
