+++ mm2src/coins/utxo/utxo_builder/mod.rs
@@ -5,5 +5,5 @@ mod utxo_conf_builder;
 pub use utxo_arc_builder::{BlockHeaderUtxoArcOps, MergeUtxoArcOps, UtxoArcBuilder};
 pub use utxo_coin_builder::{UtxoCoinBuildError, UtxoCoinBuildResult, UtxoCoinBuilder, UtxoCoinBuilderCommonOps,
                             UtxoCoinWithIguanaPrivKeyBuilder, UtxoFieldsWithHardwareWalletBuilder,
-                            UtxoFieldsWithIguanaPrivKeyBuilder, UtxoSyncStatus, UtxoSyncStatusLoopHandle};
+                            UtxoFieldsWithIguanaPrivKeyBuilder};
 pub use utxo_conf_builder::{UtxoConfBuilder, UtxoConfError, UtxoConfResult};
