+++ mm2src/coins/z_coin.rs
@@ -3,7 +3,7 @@ use crate::rpc_command::init_withdraw::{InitWithdrawCoin, WithdrawInProgressStat
 use crate::utxo::rpc_clients::{ElectrumRpcRequest, UnspentInfo, UtxoRpcClientEnum, UtxoRpcError, UtxoRpcFut,
                                UtxoRpcResult};
 use crate::utxo::utxo_builder::{UtxoCoinBuilderCommonOps, UtxoCoinWithIguanaPrivKeyBuilder,
-                                UtxoFieldsWithIguanaPrivKeyBuilder, UtxoSyncStatusLoopHandle};
+                                UtxoFieldsWithIguanaPrivKeyBuilder};
 use crate::utxo::utxo_common::{addresses_from_script, big_decimal_from_sat, big_decimal_from_sat_unsigned,
                                payment_script};
 use crate::utxo::{sat_from_big_decimal, utxo_common, ActualTxFee, AdditionalTxData, Address, BroadcastTxErr,
@@ -728,8 +728,6 @@ impl<'a> UtxoCoinBuilderCommonOps for ZCoinBuilder<'a> {
     fn activation_params(&self) -> &UtxoActivationParams { &self.utxo_params }
 
     fn ticker(&self) -> &str { self.ticker }
-
-    fn sync_status_loop_handle(&self) -> Option<UtxoSyncStatusLoopHandle> { None }
 }
 
 #[async_trait]
